gpo-assessment/requirements.txt
psycopg2-binary==2.9.10
redis==5.2.1
xlsxwriter==3.2.2
python-docx==1.1.2
